refactor: rename to microsoft-hermes, use parseArgs, add PR links
1. Rename microsoft-resolveHermes.js → microsoft-hermes.js for brevity
2. Use Node's parseArgs + switch/case instead of manual argv parsing,
   matching the pattern in change.mts
3. Add tracking PR links to the recompose JSDoc for when upstream
   Hermes includes macOS natively and the recompose can be removed
