+++ packages/react-native/scripts/ios-prebuild/hermes.js
@@ -11,7 +11,7 @@
 const {
   findMatchingHermesVersion,
   hermesCommitAtMergeBase,
-} = require('./microsoft-resolveHermes'); // [macOS]
+} = require('./microsoft-hermes'); // [macOS]
 const {computeNightlyTarballURL, createLogger} = require('./utils');
 const {execSync} = require('child_process');
 const fs = require('fs');
